fix(articles): update implementation
- update implementation
- rebase on develop

[Fixes #168064097]
