+++ src/helpers/article.js
@@ -66,17 +66,17 @@ export const getTagName = async (tags) => {
 };
 
 /**
-   * filter for search article
-   *
-   * @function
-   *
-   * @param {title} title - express request object
-   * @param {tag} tag - express request object
-   * @param {author} author - express request object
-   * @param {keyword} keyword - express request object
-   *
-   * @returns {Object} - sequlize Object
-   */
+ * filter for search article
+ *
+ * @function
+ *
+ * @param {title} title - express request object
+ * @param {tag} tag - express request object
+ * @param {author} author - express request object
+ * @param {keyword} keyword - express request object
+ *
+ * @returns {Object} - sequlize Object
+ */
 const filter = (title, tag, author, keyword) => {
   const queryParams = {
     query: {
@@ -87,8 +87,8 @@ const filter = (title, tag, author, keyword) => {
       [Op.or]: {
         '$author.firstName$': { [Op.iLike]: `%${author}%` },
         '$author.lastName$': { [Op.iLike]: `%${author}%` },
-        '$author.userName$': { [Op.iLike]: `%${author}%` }
-      }
+        '$author.userName$': { [Op.iLike]: `%${author}%` },
+      },
     },
     keyword: {
       [Op.or]: {
@@ -97,43 +97,63 @@ const filter = (title, tag, author, keyword) => {
         '$author.firstName$': { [Op.iLike]: `%${keyword}%` },
         '$author.lastName$': { [Op.iLike]: `%${keyword}%` },
         '$author.userName$': { [Op.iLike]: `%${keyword}%` },
-      }
-    }
+      },
+    },
   };
 
   if (!title) delete queryParams.query['$article.title$'];
   if (!tag) delete queryParams.query['$tag.category$'];
 
-  // eslint-disable-next-line no-nested-ternary
-  const filterQuery = (keyword) ? { ...queryParams.keyword } : author ? { ...queryParams.author }
-    : { ...queryParams.query };
+  let filterQuery;
+
+  if (keyword) {
+    filterQuery = { ...queryParams.keyword };
+  }
+  if (author) {
+    filterQuery = { ...queryParams.author };
+  } else {
+    filterQuery = { ...queryParams.query };
+  }
+
+  const hasFilterQuery = keyword || title || tag || author;
 
   return {
-    include: [{ model: Users, as: 'author' }, { model: Articles, as: 'article' }, { model: Categories, as: 'tag' }],
-    where: (keyword || title || tag || author) ? {
-      ...filterQuery
-    } : {}
+    include: [
+      { model: Users, as: 'author' },
+      { model: Articles, as: 'article' },
+      { model: Categories, as: 'tag' },
+    ],
+    where: hasFilterQuery && { ...filterQuery }
   };
 };
 
 /**
-     * function for extract article
-     *
-     * @function
-     *
-     * @param {result} result - array of object
-     *
-     * @returns {Object} - callback that execute the controller
-     */
-const extractArticles = result => result.map((entry) => {
+ * function for extract article
+ *
+ * @function
+ *
+ * @param {result} result - array of object
+ *
+ * @returns {Object} - callback that execute the controller
+ */
+const extractArticles = result => result.reduce((accumulator, entry) => {
   const firstName = entry['author.firstName'];
   const lastName = entry['author.lastName'];
   const userName = entry['author.userName'];
   const createdAt = entry['author.createdAt'];
-  return {
-    ArticleId: entry['article.id'],
+  const ArticleId = entry['article.id'];
+  const item = accumulator.find(value => value.ArticleId === entry.articleId);
+
+  if (item) {
+    if (item.category.includes(entry['tag.category'])) return accumulator;
+    item.category.push(entry['tag.category']);
+
+    return accumulator;
+  }
+
+  const newEntry = {
+    ArticleId,
     title: entry['article.title'],
-    tag: entry['tag.category'],
     summary: entry['article.summary'],
     body: entry['article.body'],
     suspended: entry['article.suspended'],
@@ -142,9 +162,12 @@ const extractArticles = result => result.map((entry) => {
     followUpId: entry['article.followUpId'],
     createdAt: entry['article.createdAt'],
     author: { name: `${firstName} ${lastName}`, userName, createdAt },
-    tags: entry.tag,
+    category: [entry['tag.category']],
   };
-});
+  accumulator[accumulator.length] = newEntry;
+
+  return accumulator;
+}, []);
 
 export default {
   idPresent,
